+++ src/workflows/zowe/workflows_for_zowe_sdk/response/__init__.py
@@ -1,0 +1,55 @@
+"""Zowe Client Python SDK.
+
+This program and the accompanying materials are made available under the terms of the
+Eclipse Public License v2.0 which accompanies this distribution, and is available at
+
+https://www.eclipse.org/legal/epl-v20.html
+
+SPDX-License-Identifier: EPL-2.0
+
+Copyright Contributors to the Zowe Project.
+"""
+
+from .workflows import (
+    ArchivedWorkflowStepResponse,
+    CreateWorkflowResponse,
+    GetArchivedWorkflowPropertiesResponse,
+    GetWorkflowDefinitionResponse,
+    GetWorkflowPropertiesResponse,
+    ListArchivedWorkflowsResponse,
+    ListWorkflowsResponse,
+    WorkflowAutomationStatusResponse,
+    WorkflowDefinitionApproverResponse,
+    WorkflowDefinitionStepResponse,
+    WorkflowDefinitionVariableResponse,
+    WorkflowJobFileResponse,
+    WorkflowJobInfoResponse,
+    WorkflowJobStatusResponse,
+    WorkflowPropertyMappingResponse,
+    WorkflowStepResponse,
+    WorkflowVariableReferenceResponse,
+    WorkflowVariableResponse,
+    WorkflowVariableSpecificationResponse,
+)
+
+__all__ = [
+    "ArchivedWorkflowStepResponse",
+    "CreateWorkflowResponse",
+    "GetArchivedWorkflowPropertiesResponse",
+    "GetWorkflowDefinitionResponse",
+    "GetWorkflowPropertiesResponse",
+    "ListArchivedWorkflowsResponse",
+    "ListWorkflowsResponse",
+    "WorkflowAutomationStatusResponse",
+    "WorkflowDefinitionApproverResponse",
+    "WorkflowDefinitionStepResponse",
+    "WorkflowDefinitionVariableResponse",
+    "WorkflowJobFileResponse",
+    "WorkflowJobInfoResponse",
+    "WorkflowJobStatusResponse",
+    "WorkflowPropertyMappingResponse",
+    "WorkflowStepResponse",
+    "WorkflowVariableReferenceResponse",
+    "WorkflowVariableResponse",
+    "WorkflowVariableSpecificationResponse",
+]